Reorganized functions within server_setup and server_config
