+++ merlin/main.py
@@ -54,10 +54,9 @@ from merlin.examples.generator import list_examples, setup_example
 from merlin.log_formatter import setup_logging
 from merlin.server.server_setup import (
     ServerStatus,
-    create_server_config,
-    get_server_status,
-    pull_server_image,
+    init_server,
     start_server,
+    status_server,
     stop_server,
 )
 from merlin.spec.expansion import RESERVED, get_spec_with_expansion
@@ -352,28 +351,13 @@ def process_monitor(args):
 
 def process_server(args):
     if args.commands == "init":
-        if not create_server_config():
-            LOG.info("Merlin server initialization failed.")
-            return
-        if pull_server_image():
-            LOG.info("New merlin server image fetched")
-        LOG.info("Merlin server initialization successful.")
+        init_server()
     elif args.commands == "start":
         start_server()
     elif args.commands == "stop":
         stop_server()
     elif args.commands == "status":
-        current_status = get_server_status()
-        if current_status == ServerStatus.NOT_INITALIZED:
-            LOG.info("Merlin server has not been initialized.")
-            LOG.info("Please initalize server by running 'merlin server init'")
-        elif current_status == ServerStatus.MISSING_CONTAINER:
-            LOG.info("Unable to find server image.")
-            LOG.info("Ensure there is a .sif file in merlin server directory.")
-        elif current_status == ServerStatus.NOT_RUNNING:
-            LOG.info("Merlin server is not running.")
-        elif current_status == ServerStatus.RUNNING:
-            LOG.info("Merlin server is running.")
+        status_server()
 
 
 def setup_argparse() -> None:
