+++ merlin/server/merlin_server.yaml
@@ -11,6 +11,10 @@ container:
     config_dir: merlin_server/
     # Process file containing information regarding the redis process
     pfile: merlin_server.pf
+    # Password file to be used for accessing container
+    pass_file: redis.pass
+    # Password command for generating password file
+    pass_command: date +%s | sha256sum
 
 process:
     # Command for determining the process of the command
